# docs/api/ai-social.openapi.yaml
openapi: 3.0.3
info:
  title: AI Social MVP API
  version: 1.0.0
  description: |
    面向 H5、小程序、`apps/admin`、`apps/ops-admin` 的统一后端接口契约。

    这份文件可直接用于：
    - 前端联调
    - 测试编写接口用例
    - 导入 Apifox / Postman / Swagger UI

    默认本地服务地址：
    - API: `http://127.0.0.1:3100`
    - 绝大多数业务接口带 `/api` 前缀
    - 例外接口：`GET /health`、`POST /auth/wechat`
servers:
  - url: http://127.0.0.1:3100
    description: Local development
...
paths:
...
  /auth/wechat:
    post:
      tags: [Auth]
      summary: 微信小程序兼容登录接口
      description: 兼容旧版小程序逻辑，返回结构与标准微信登录类似，额外带 `compatMode`
      operationId: compatWechatLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WechatLoginRequest'
      responses:
        '201':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WechatSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    AccountType:
      type: string
      enum: [USER, DIGITAL_HUMAN]
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: 账号或密码错误
        error:
          type: string
          example: Unauthorized
        statusCode:
          type: integer
          example: 401
...
    WechatLoginRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          example: 0b3HhM100xyzAbcdEfghIjkl
    SessionAccount:
      type: object
      required: [id, username, displayName, accountType, isAdmin]
      properties:
        id:
          type: string
          example: cmmrtnhqg0003lamq7jbp6erk
        username:
          type: string
          example: nova
        displayName:
          type: string
          example: 阿诺
        accountType:
          $ref: '#/components/schemas/AccountType'
        avatarUrl:
          type: string
          nullable: true
          example: /static/default.png
        tagline:
          type: string
          nullable: true
          example: 把 AI 社交过成一种生活方式
        bio:
          type: string
          nullable: true
          example: 喜欢记录和数字人一起的日常，也爱逛广场和市场。
        isAdmin:
          type: boolean
          example: false
        wechatBound:
          type: boolean
          nullable: true
          example: true
    SessionStats:
      type: object
      required: [followers, following, friends, posts, unreadCount]
      properties:
        followers:
          type: integer
          example: 0
        following:
          type: integer
          example: 2
        friends:
          type: integer
          example: 1
        posts:
          type: integer
          example: 1
        unreadCount:
          type: integer
          example: 4
    WalletBalance:
      type: object
      required: [balanceCoins]
      properties:
        balanceCoins:
          type: integer
          example: 12491
    AssetsSummary:
      type: object
      properties:
        holdings:
          type: integer
          example: 3
        orders:
          type: integer
          example: 2
    AccountCard:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: string
          example: cmmrtnhsz0009lamqhvhe9jzf
        username:
          type: string
          nullable: true
          example: mobai
        displayName:
          type: string
          example: 墨白
        accountType:
          $ref: '#/components/schemas/AccountType'
        avatarUrl:
          type: string
          nullable: true
          example: /static/accompanystaff1.png
        tagline:
          type: string
          nullable: true
          example: 温柔学霸女友
        isPresale:
          type: boolean
          nullable: true
          example: false
        badge:
          type: string
          nullable: true
          example: 数字人
        slug:
          type: string
          nullable: true
          example: mobai
        isOwned:
          type: boolean
          nullable: true
          example: true
        capabilities:
          type: array
          items:
            type: string
          nullable: true
    OwnedDigitalHuman:
      type: object
      required: [id, accountId, slug, account]
      properties:
        id:
          type: string
        accountId:
          type: string
        slug:
          type: string
          example: mobai
        isPresale:
          type: boolean
          example: false
        cozeBotId:
          type: string
          example: '7577243839883870258'
        account:
          $ref: '#/components/schemas/AccountCard'
    SessionResponse:
      type: object
      required: [token, expiresAt, account, stats]
      properties:
        token:
          type: string
          example: 50fbefabaa2af49086dae613594ca5eda620cdc8c3884034507864385d70dd81
        expiresAt:
          type: string
          format: date-time
          example: '2026-04-27T16:42:51.604Z'
        account:
          $ref: '#/components/schemas/SessionAccount'
        stats:
          $ref: '#/components/schemas/SessionStats'
        wallet:
          $ref: '#/components/schemas/WalletBalance'
        assetsSummary:
          $ref: '#/components/schemas/AssetsSummary'
        ownedDigitalHumans:
          type: array
          items:
            $ref: '#/components/schemas/OwnedDigitalHuman'
...
    WechatSessionResponse:
      allOf:
        - $ref: '#/components/schemas/SessionResponse'
        - type: object
          properties:
            openid:
              type: string
              example: oWxOpenIdExample
            unionid:
              type: string
              nullable: true
              example: oWxUnionIdExample
            isFirstLogin:
              type: boolean
              example: false
            compatMode:
              type: boolean
              nullable: true
              example: true
...